КР 2

+++ Разборы/КР 2/a.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-    int n, a, b;
+    long long n, a, b;
     cin >> n >> b >> a;
 
     for(int i = 0; i < n; i++) {
